Composite v5 CONNECT request
* Extract `write_request`, `read_request_status` and associated types into module `client_commands`.

* Add `request_connect` and `SocksV5ConnectError`.
