+++ src/v5/mod.rs
@@ -5,11 +5,13 @@ use thiserror::Error;
 #[cfg(feature = "tokio")]
 use tokio_compat::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 
+mod client_commands;
 mod handshake;
 mod hosts;
 mod types;
 
 use crate::SocksVersion;
+pub use client_commands::*;
 pub use handshake::*;
 pub use hosts::*;
 pub use types::*;
@@ -107,142 +109,6 @@ where
     })
 }
 
-/// Writes a SOCKSv5 request with the specified command, host and port.
-///
-/// # Errors
-///
-/// If writing to `writer` fails, this function will return the I/O error.
-///
-/// # Panics
-///
-/// If `host` is a domain name, the length of which is greater than 255 bytes.
-/// The SOCKSv5 specification leaves only a single octet for encoding the domain name length,
-/// so a target longer than 255 bytes cannot be properly encoded.
-pub async fn write_request<Writer>(
-    mut writer: Writer,
-    command: SocksV5Command,
-    host: SocksV5Host,
-    port: u16,
-) -> std::io::Result<()>
-where
-    Writer: AsyncWrite + Unpin,
-{
-    let mut data = Vec::<u8>::with_capacity(
-        6 + match &host {
-            SocksV5Host::Domain(domain) => {
-                assert!(
-                    domain.len() <= 256,
-                    "domain name must be shorter than 256 bytes"
-                );
-                1 + domain.len()
-            }
-            SocksV5Host::Ipv4(_) => 4,
-            SocksV5Host::Ipv6(_) => 16,
-        },
-    );
-    data.push(SocksVersion::V5.to_u8());
-    data.push(command.to_u8());
-    data.push(0u8); // reserved bits in SOCKSv5
-    match &host {
-        SocksV5Host::Domain(domain) => {
-            data.push(SocksV5AddressType::Domain.to_u8());
-            data.push(domain.len() as u8);
-            data.extend_from_slice(domain);
-        }
-        SocksV5Host::Ipv4(octets) => {
-            data.push(SocksV5AddressType::Ipv4.to_u8());
-            data.extend_from_slice(octets);
-        }
-        SocksV5Host::Ipv6(octets) => {
-            data.push(SocksV5AddressType::Ipv6.to_u8());
-            data.extend_from_slice(octets);
-        }
-    }
-    let port_start = data.len();
-    data.extend_from_slice(b"\0\0");
-    NetworkEndian::write_u16(&mut data[port_start..], port);
-
-    writer.write_all(&data).await
-}
-
-#[derive(Debug)]
-pub struct SocksV5Response {
-    pub status: SocksV5RequestStatus,
-    pub host: SocksV5Host,
-    pub port: u16,
-}
-
-pub type SocksV5ResponseResult = Result<SocksV5Response, SocksV5RequestError>;
-
-/// Reads and parses a SOCKSv5 command response message.
-///
-/// Depending on the data (in case of parsing errors),
-/// this function may not consume the whole response from the server.
-///
-/// # Errors
-///
-/// If reading from `reader` fails, including if a premature EOF is encountered,
-/// this function will return the I/O error (wrapped in `SocksV5RequestError::Io`).
-///
-/// If the first byte read from `reader` is not `05`, as required by the SOCKSv5 specification,
-/// then this function will return `SocksV5RequestError::InvalidVersion` with the actual "version number".
-///
-/// If the status byte or the address type byte are not from the respective lists in the specification,
-/// then this function will return `SocksV5RequestError::InvalidRequest`
-/// with a human-readable description of the error.
-pub async fn read_request_status<Reader>(mut reader: Reader) -> SocksV5ResponseResult
-where
-    Reader: AsyncRead + Unpin,
-{
-    let mut buf = [0u8; 2];
-
-    reader.read_exact(&mut buf[0..1]).await?;
-    if buf[0] != SocksVersion::V5.to_u8() {
-        return Err(SocksV5RequestError::InvalidVersion(buf[0]));
-    }
-
-    reader.read_exact(&mut buf[0..1]).await?;
-    let status = SocksV5RequestStatus::from_u8(buf[0]).ok_or_else(|| {
-        SocksV5RequestError::InvalidRequest(format!("invalid status {:02X}", buf[0]))
-    })?;
-
-    reader.read_exact(&mut buf).await?;
-    // ignore a reserved octet, use the following one
-    let atyp = SocksV5AddressType::from_u8(buf[1]).ok_or_else(|| {
-        SocksV5RequestError::InvalidRequest(format!(
-            "invalid address type {:02X}, expected {:02X} (IP V4), {:02X} (DOMAINNAME), or {:02X} (IP V6)",
-            buf[1],
-            SocksV5AddressType::Ipv4.to_u8(),
-            SocksV5AddressType::Domain.to_u8(),
-            SocksV5AddressType::Ipv6.to_u8(),
-        ))
-    })?;
-
-    let host = match atyp {
-        SocksV5AddressType::Ipv4 => {
-            let mut host = [0u8; 4];
-            reader.read_exact(&mut host).await?;
-            SocksV5Host::Ipv4(host)
-        }
-        SocksV5AddressType::Ipv6 => {
-            let mut host = [0u8; 16];
-            reader.read_exact(&mut host).await?;
-            SocksV5Host::Ipv6(host)
-        }
-        SocksV5AddressType::Domain => {
-            reader.read_exact(&mut buf[0..1]).await?;
-            let mut domain = vec![0u8; buf[0] as usize];
-            reader.read_exact(&mut domain).await?;
-            SocksV5Host::Domain(domain)
-        }
-    };
-
-    reader.read_exact(&mut buf).await?;
-    let port = NetworkEndian::read_u16(&buf);
-
-    Ok(SocksV5Response { status, port, host })
-}
-
 pub async fn write_request_status<Writer>(
     mut writer: Writer,
     status: SocksV5RequestStatus,
@@ -290,35 +156,6 @@ mod tests {
     use futures::executor::block_on;
 
     use super::*;
-
-    #[test]
-    fn write_request_ipv4() {
-        let mut buf = Vec::<u8>::new();
-        block_on(write_request(
-            &mut buf,
-            SocksV5Command::Connect,
-            SocksV5Host::Ipv4([127, 0, 0, 1]),
-            1080,
-        ))
-        .unwrap();
-        assert_eq!(buf, &[5, 1, 0, 1, 127, 0, 0, 1, 4, 56]);
-    }
-
-    #[test]
-    fn write_request_ipv6() {
-        let mut buf = Vec::<u8>::new();
-        block_on(write_request(
-            &mut buf,
-            SocksV5Command::Connect,
-            SocksV5Host::Ipv6([1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6]),
-            1080,
-        ))
-        .unwrap();
-        assert_eq!(
-            buf,
-            &[5, 1, 0, 4, 1, 2, 3, 4, 5, 6, 7, 8, 9, 0, 1, 2, 3, 4, 5, 6, 4, 56]
-        );
-    }
 
     #[test]
     fn write_request_domain() {
